Move LoadAsDirectoryAsync to new helper class SteamDirectory, wrapping WebAPI.
